new/pyproject.toml
[project]
name = "new"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.129.0",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.41.0",
]
